refactor: move preflight scoring and utilities
