+++ tldw_Server_API/app/core/Web_Scraping/preflight/recommendations/__init__.py
@@ -1,0 +1,7 @@
+"""Recommendation helpers based on aggregated analyzer results."""
+
+from __future__ import annotations
+
+from .recommender import generate_recommendations
+
+__all__ = ["generate_recommendations"]
