+++ tldw_Server_API/app/core/Web_Scraping/preflight/scoring/__init__.py
@@ -1,0 +1,7 @@
+"""Difficulty scoring helpers for scraping analysis."""
+
+from __future__ import annotations
+
+from .scoring_engine import calculate_difficulty_score
+
+__all__ = ["calculate_difficulty_score"]
