+++ tldw_Server_API/app/core/Web_Scraping/scraper_analyzers/recommendations/__init__.py
@@ -1,7 +1,9 @@
-"""Recommendation helpers based on aggregated analyzer results."""
+"""Deprecated compatibility re-export; implementation lives in Web_Scraping.preflight."""
 
 from __future__ import annotations
 
-from .recommender import generate_recommendations
+from tldw_Server_API.app.core.Web_Scraping.preflight.recommendations.recommender import (
+    generate_recommendations,
+)
 
 __all__ = ["generate_recommendations"]
