+++ tldw_Server_API/app/core/Web_Scraping/scraper_analyzers/scoring/__init__.py
@@ -1,7 +1,9 @@
-"""Difficulty scoring helpers for scraping analysis."""
+"""Deprecated compatibility re-export; implementation lives in Web_Scraping.preflight."""
 
 from __future__ import annotations
 
-from .scoring_engine import calculate_difficulty_score
+from tldw_Server_API.app.core.Web_Scraping.preflight.scoring.scoring_engine import (
+    calculate_difficulty_score,
+)
 
 __all__ = ["calculate_difficulty_score"]
